perf(zkevm): run ParallelUnbalancedWork loops inline in the guest
The guest has no threads, yet the Trie's branch RLP helper still went
through ParallelUnbalancedWork.For, so the image carried the work-item
class, its shared counters and the thread-pool and Task code they pull
in for loops that never fan out.

The class is now split the way VirtualMachine is: the shared file keeps
DefaultOptions and the public overloads, ParallelUnbalancedWork.std.cs
keeps the threaded implementation behind two private partial methods,
and ParallelUnbalancedWork.zkevm.cs runs the range in order on the
calling thread under the same cancellation and local-state contract.
Call sites do not change.
